+++ api/test/metrics/meter_provider_test.cc
@@ -1,12 +1,10 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef ENABLE_METRICS_PREVIEW
-
-#  include <gtest/gtest.h>
-#  include "opentelemetry/metrics/noop.h"
-#  include "opentelemetry/metrics/provider.h"
-#  include "opentelemetry/nostd/shared_ptr.h"
+#include <gtest/gtest.h>
+#include "opentelemetry/metrics/noop.h"
+#include "opentelemetry/metrics/provider.h"
+#include "opentelemetry/nostd/shared_ptr.h"
 
 using opentelemetry::metrics::Meter;
 using opentelemetry::metrics::MeterProvider;
@@ -35,4 +33,3 @@ TEST(Provider, MultipleMeterProviders)
 
   ASSERT_NE(Provider::GetMeterProvider(), tf);
 }
-#endif