+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_grpc_log_record_exporter_factory.h
@@ -1,0 +1,39 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#ifdef ENABLE_LOGS_PREVIEW
+
+#  include "opentelemetry/exporters/otlp/otlp_grpc_exporter_options.h"
+#  include "opentelemetry/sdk/logs/exporter.h"
+
+OPENTELEMETRY_BEGIN_NAMESPACE
+namespace exporter
+{
+namespace otlp
+{
+
+/**
+ * Factory class for OtlpGrpcLogRecordExporter.
+ */
+class OtlpGrpcLogRecordExporterFactory
+{
+public:
+  /**
+   * Create a OtlpGrpcLogRecordExporter.
+   */
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create();
+
+  /**
+   * Create a OtlpGrpcLogRecordExporter.
+   */
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create(
+      const OtlpGrpcExporterOptions &options);
+};
+
+}  // namespace otlp
+}  // namespace exporter
+OPENTELEMETRY_END_NAMESPACE
+
+#endif /* ENABLE_LOGS_PREVIEW */