+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_http_log_record_exporter_factory.h
@@ -1,0 +1,39 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#ifdef ENABLE_LOGS_PREVIEW
+
+#  include "opentelemetry/exporters/otlp/otlp_http_log_record_exporter_options.h"
+#  include "opentelemetry/sdk/logs/exporter.h"
+
+#  include <memory>
+
+OPENTELEMETRY_BEGIN_NAMESPACE
+namespace exporter
+{
+namespace otlp
+{
+
+/**
+ * Factory class for OtlpHttpLogRecordExporter.
+ */
+class OtlpHttpLogRecordExporterFactory
+{
+public:
+  /**
+   * Create a OtlpHttpLogRecordExporter.
+   */
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create();
+
+  /**
+   * Create a OtlpHttpLogRecordExporter.
+   */
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create(
+      const OtlpHttpLogRecordExporterOptions &options);
+};
+
+}  // namespace otlp
+}  // namespace exporter
+OPENTELEMETRY_END_NAMESPACE
+#endif