+++ sdk/src/metrics/exemplar/filter.cc
@@ -1,11 +1,10 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef ENABLE_METRICS_PREVIEW
-#  include "opentelemetry/sdk/metrics/exemplar/filter.h"
-#  include "opentelemetry/sdk/metrics/exemplar/always_sample_filter.h"
-#  include "opentelemetry/sdk/metrics/exemplar/never_sample_filter.h"
-#  include "opentelemetry/sdk/metrics/exemplar/with_trace_sample_filter.h"
+#include "opentelemetry/sdk/metrics/exemplar/filter.h"
+#include "opentelemetry/sdk/metrics/exemplar/always_sample_filter.h"
+#include "opentelemetry/sdk/metrics/exemplar/never_sample_filter.h"
+#include "opentelemetry/sdk/metrics/exemplar/with_trace_sample_filter.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace sdk
@@ -34,4 +33,3 @@ std::shared_ptr<ExemplarFilter> ExemplarFilter::GetWithTraceSampleFilter() noexc
 }  // namespace metrics
 }  // namespace sdk
 OPENTELEMETRY_END_NAMESPACE
-#endif