+++ sdk/test/metrics/exemplar/reservoir_cell_test.cc
@@ -1,9 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef ENABLE_METRICS_PREVIEW
-#  include "opentelemetry/sdk/metrics/exemplar/reservoir_cell.h"
-#  include <gtest/gtest.h>
+#include "opentelemetry/sdk/metrics/exemplar/reservoir_cell.h"
+#include <gtest/gtest.h>
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace sdk
@@ -71,4 +70,3 @@ TEST_F(ReservoirCellTestPeer, Filtered)
 }  // namespace metrics
 }  // namespace sdk
 OPENTELEMETRY_END_NAMESPACE
-#endif